requirements.txt
openai-whisper
python-docx
tqdm
reportlab